Completion hacks W.I.P.
I discovered a little gem in bsh.util.NameCompletionTable it's the guy that does the JConsole completion.
Gets bootstrapped with the interpreter namespace and can also work the classpath. Do try with JDK8 with rt.jar loaded for pretty neat insights, Java 9 currently can't get that detail.
The whole MappingFeedback listener is a tad odd, without it the ClassPath echo's to stderr what it's doing.
The rest is an attempt to dive deeper obtaining members for completion, it uses bsh.Name (nameResolver)'s magic which can be a bit like a magicians top hat, you never know what you're going to get. Good proof of concept me thinks.